#25 Test [ci full] option
* Improve execution time and fix warnings

* Merge with master

* Add [ci full] option

* Use travis branch

* Test commit with [ci full] message

* Test travis env vars

* Test commit msg from bash

* Check diff level of git log

* Add TRAVIS_COMMIT_MESSAGE env var

* Add echos to preinstall

* Test Add echos to preinstall

* Implement [ci full] option

* Add logging to preinstall

* Export TRAVIS_COMMIT_MESSAGE

* Chech [ci full] option in commit msg

* Remove logging

+++ bin/test/install.sh
@@ -36,22 +36,38 @@ if [ "$__IS_CONCURRENT_SCRIPT" == "$__NONE" ] || [ "$__IS_CONCURRENT_SCRIPT" == 
 
   echo "Start initializing backend"
 
-  ######################################################
-  ### TODO: Check if file exists, otherwise exit (?) ###
-  ######################################################
+  ###########################################################################
+  ### Check if deeploy.example.json exists, otherwise exit with error msg ###
+  ###########################################################################
+  if [ ! -f ${__SRC_PATH}deeploy.example.json ]; then
+    echo "File deeploy.example.json doesn't exist. Please create it and repeat"
+    exit 1
+  fi
+
   cp ${__SRC_PATH}deeploy.example.json ${__SRC_PATH}deeploy.json
 
   if [ "${TRAVIS}" == "true" ] && [ "$BACKEND_MICROAPP_IDENTIFIERS" == "none" ]; then
+
+    ###########################################################################
+    ### Skip initializing backend if no changes in backend or running in CI ###
+    ###########################################################################
     echo "Skipping initializing backend, becuase no changes in backend"
   elif [ "$BACKEND_MICROAPP_IDENTIFIERS" == "$__NONE" ] || [ -z "$BACKEND_MICROAPP_IDENTIFIERS" ] || \
     ( ([ "BACKEND_MICROAPP_PATHS" != "$__NONE" ] || [ "FRONTEND_MICROAPP_PATHS" != "$__NONE" ]) && \
     ([ "${__E2E_WITH_PUBLIC_REPO}" == "${E2E_TESTING}" ] || ([ "${__E2E_WITH_PRIVATE_REPO}" == "${E2E_TESTING}" ] && \
     [ "${TRAVIS_BRANCH}" == 'stage' ]))); then
 
+    #################################################################
+    ### Fully initializing backend for e2e test or forced locally ###
+    ##########################################################
     echo "Fully initializing backend"
 
     cd ${__SRC_PATH} && deepify compile dev "${__SRC_PATH}"
   else
+
+    ################################################################
+    ### Partially initializing backend for specified identifiers ###
+    ################################################################
     echo "Partially initializing backend: ${BACKEND_MICROAPP_IDENTIFIERS}"
     cd ${__SRC_PATH} && deepify compile dev -m "${BACKEND_MICROAPP_IDENTIFIERS}"
   fi
